+++ src/app/entities/RaceHassouJoukyou.ts
@@ -11,7 +11,7 @@ export class RaceHassouJoukyou {
   @PrimaryColumn("bigint")
   public Id: number;
 
-  @Column("bigint")
+  @Column("int")
   public RaceId: number;
 
   @Column("varchar", { length: 120 })
