+++ src/app/entities/RaceLapTime.ts
@@ -11,7 +11,7 @@ export class RaceLapTime {
   @PrimaryColumn("bigint")
   public Id: number;
 
-  @Column("bigint")
+  @Column("int")
   public RaceId: number;
 
   @Column("smallint")
